hosts/zeus: Refactor and Fix the iscsi integration
The iSCSI integration I was using was made up from a bunch of hacks and
try-error and it was never working properly. When I changed to LACP it
completely broke. This change gets rid of all the services (hacks), and
replace them with properly written services, copied from the open-iscsi
project over on GitHub[0].

[0]: https://github.com/open-iscsi/open-iscsi/tree/f37d5b653f9f251845db3f29b1a3dcb90ec89731/etc/systemd