hosts/zeus: re-enable firewall
It turns out that the firewall works fine but only when I'm logging in
from the admin interface to the admin interface. It does not seem to
allow me to go from the general interface to the admin interface. But
that's OK! I'm going to block access to the admin interface from
anywhere else through my router anyway...