profiles/{myself,work/keeptruckin}: Include SSH config by HOME path
Eternal-Terminal does not currently behave in the same way SSH does with
regards to an include statement with a relative path.

Excerpt from `man ssh_config':

	Files without absolute paths are assumed to be in ~/.ssh if included in a user configuration file.

I filed a [bug upstream][bug], but as a workaround I've changed the include
statements to absolute paths in the format `~/.ssh/<relative path>`.

[bug]: https://github.com/MisterTea/EternalTerminal/issues/371