modules/programs/weechat: new module enabled on workstations
Copied the configuration from [cole-h][cole-h], and updated the config
to use read-only files.

I've also provided an 'edit-weechat-config' to open weechat with the
configuration copied to a temporary location, and the encrypted files
decrypted to allow the user to edit the config. Once weechat exits, the
configuration is then copied back, and the encrypted files re-encrypted.

[cole-h]: https://github.com/cole-h/nixos-config/blob/92749c8d5fa807692298c20c3819df07774c235c/users/vin/modules/weechat/default.nix