Add support for direct home-manager hosts (#14)
We currently only support a NixOS system, this adds the ability to create a home-manager-based configuration that works for any system. I've only tested Debian.

There are many other changes in this PR that includes an update to the latest nixpkgs/home-manager, and moving things around, and refactor code. Too lazy to split them.