profiles/core: configure my keyboard to accommodate ErgoDox EZ
ErgoDox has its own way of configuring the layout so long as the
software configures it as US layout. In other words, if I configured the
keyboard to the Colemak layout and done the same to my OS then the
keyboard comes out as an undefined layout (neither Colemak nor US).

I tried setting up Colemak being the first layout so it remains the
default layout and then add a US layout and for some reason i3 started
looking for Mod4+j to move focus left even though in the configuration
it's clearly configured as Mod4+n

I tried setting up Colemak being the first layout so it remains the
default layout and then add a US layout and for some reason i3 started
looking for Mod4+j to move focus left even though in the configuration
it's clearly configured as Mod4+n.

As a workaround, I configured the layouts to be US first then Colemak
but override the defaultLayout to Colemak. With this hack the
encryption password (early console) and the console remains configured
to Colemak but the default layout in i3 is Colemak. I've also configured
Right Ctrl + Right Shift to switch between layouts so I can use both of
my external and internal keyboards.