programs.fzf: init module to extend fzf configuration
I removed programs.fzf.defaultCommand in #17 because I thought it would
be better to navigate all the files not just the ones that are tracked
by Git. However, I keep matching files inside directories that I
explicitly ignored with Git.

I do want to be able to navigate all files, but I'd probably end up
using a separate Vim binding for that, but until that I'm going to
revert this change.