profiles: setup AWS credentials at runtime
At work, AWS authentication happens with a script that uses Okta to
login. The credentials are only temporary, so the configuration files
must be writable and not tracked by Soxin.

By using a shell script to use `aws configure` to set values upon home
activation, the AWS files becomes writable and owned by the user.