profiles/workstation: generate a new GnuPG ECC key for Onlykey
Unfortunately I was not able to create a GnuPG key with the root only
certifying and two sub-keys for decryption and signing because it's not
supported. I kept getting an error (in the agent itself) that the
key ID for the signing key is not the same as the ID of the root key.