programs/zsh: Fix Keychain search for pubkey
The Keychain program's logic does not make sense.

My setup is as follows:
- `~/.ssh/id_ed25519_sk_rk` is a private key, a symlink into a secret location.
- `~/.ssh/id_ed25519_sk_rk.pub` is a public key, also a symlink but somewhere inside the Nix Store.

Keychain wants to find the signature of the SSH key in the `ssh_f()` function, and instead of testing
the presence of the files above, it gets the absolute path of `~/.ssh/id_ed25519_sk_rk` and looks for
a .pub with the same name in my secret location. I mean come one, Keychain, it's literally RIGHT THERE!