services/{sketchybar,skhd,yabai}: Remove auto-install/restart due to Segfault
I suspect this is the same issue that led to these commits:
- c24d3351cf0cbdae5f4b2b9f3e899b212add6840
- 334887d3e5c97c3dd01b3260e64983aa1b512d36
- d26f5b91a4c8853f876924bc12758d7095228f04

For some reason, when launchctl is invoked, Nix crashes and sometimes
taking the Nix store with it (as in it becomes invalid and corrupt).
However, it did not corrupt the store this time, albeit, I tried only
once.