scripts/host: set the profile correctly and improve activation
The activation was mimicking what the darwin-rebuild was doing, but
actually, I should have been calling darwin-rebuild with the 'activate'
command. It does everything correctly for setting up the system and the
user, the only thing it does not do is to set up the profile which can
only be activated with a switch, and switch wants to build a flake
instead of using the already built output.