treewide: Move my home to /home/yl
It seems that Snap does not allow user's home to be outside /home (see
the forum link below). When I initially moved to `/yl` I wanted an
easier way to refer to my home but tbh most of the time I use `~/`
anyways and since it's causing problems now, I decided to revert back to
standard home directory location.

https://forum.snapcraft.io/t/11209